requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
frozenlist==1.3.3
google-auth==2.16.2
google-auth-oauthlib==1.0.0
gspread==5.7.2
gspread-asyncio==1.8.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
oauthlib==3.2.2
pillow==10.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
StrEnum==0.4.15
urllib3==1.26.14
yarl==1.8.2
